+++ src/Broadway/Bundle/BroadwayBundle/Resources/config/services.xml
@@ -23,7 +23,7 @@
 
         <service id="broadway.event_handling.event_bus" class="Broadway\EventHandling\SimpleEventBus"  lazy="true"/>
 
-        <service id="broadway.uuid.generator" class="Broadway\Uuid\Rfc4122\Version4Generator" />
+        <service id="broadway.uuid.generator" class="Broadway\UuidGenerator\Rfc4122\Version4Generator" />
 
         <service id="broadway.metadata_enriching_event_stream_decorator" class="Broadway\EventSourcing\MetadataEnrichment\MetadataEnrichingEventStreamDecorator" />
