--- src/Broadway/Uuid/README.md
@@ -1,17 +1,0 @@
-Uuid Component
-==============
-
-Uuid generators for your application.
-
-Provides a uuid generator with help of [rhumsaa/uuid] and several testing
-helpers to use uuids in your tests.
-
-[rhumsaa/uuid]: https://github.com/ramsey/uuid
-
-## Example
-
-The [`examples/`][examples] directory at the root of the project contains a
-runnable uuid example (`uuid`). The code you find there contains comments with
-what is happening.
-
-[examples]: ../../../examples/
