--- src/Broadway/Uuid/UuidGeneratorInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-/*
- * This file is part of the broadway/broadway package.
- *
- * (c) Qandidate.com <opensource@qandidate.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Broadway\Uuid;
-
-/**
- * Generates uuids.
- */
-interface UuidGeneratorInterface
-{
-    /**
-     * @return string
-     */
-    public function generate();
-}
