+++ test/Broadway/Saga/MultipleSagaManagerTest.php
@@ -22,7 +22,7 @@ use Broadway\Saga\State\InMemoryRepository;
 use Broadway\Saga\State\RepositoryInterface;
 use Broadway\Saga\State\StateManager;
 use Broadway\TestCase;
-use Broadway\Uuid\Rfc4122\Version4Generator;
+use Broadway\UuidGenerator\Rfc4122\Version4Generator;
 
 class MultipleSagaManagerTest extends TestCase
 {
